backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.1

# Image handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8

# API filtering
django-filter==24.1

# API documentation
drf-spectacular==0.27.1

# Payment
razorpay==1.4.1

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0

# Development
black==24.1.1
flake8==7.0.0
ipython==8.20.0
